+++ linphone-app/assets/languages/en.ts
@@ -3123,6 +3123,11 @@ Click here: &lt;a href=&quot;%1&quot;&gt;%1&lt;/a&gt;
         <extracomment>&apos;Enable screenshots&apos; : Label on switch to choose if we can take screenshots while being in call.</extracomment>
         <translation>Enable screenshots</translation>
     </message>
+    <message>
+        <source>chatReactionsNotificationsEnabledLabel</source>
+        <extracomment>&apos;Enable notifications on reactions&apos;: settings label for enabling notifications on chat reactions.</extracomment>
+        <translation>Enable notifications on reactions</translation>
+    </message>
 </context>
 <context>
     <name>SettingsLdap</name>