+++ linphone-app/src/components/settings/SettingsModel.hpp
@@ -173,11 +173,12 @@ class SettingsModel : public QObject {
 
 	Q_PROPERTY(bool chatNotificationsEnabled READ getChatNotificationsEnabled WRITE setChatNotificationsEnabled NOTIFY
 	               chatNotificationsEnabledChanged)
+	Q_PROPERTY(bool chatReactionsNotificationsEnabled READ getChatReactionsNotificationsEnabled WRITE
+	               setChatReactionsNotificationsEnabled NOTIFY chatReactionsNotificationsEnabledChanged)
 	Q_PROPERTY(bool chatNotificationSoundEnabled READ getChatNotificationSoundEnabled WRITE
 	               setChatNotificationSoundEnabled NOTIFY chatNotificationSoundEnabledChanged)
 	Q_PROPERTY(QString chatNotificationSoundPath READ getChatNotificationSoundPath WRITE setChatNotificationSoundPath
 	               NOTIFY chatNotificationSoundPathChanged)
-
 	Q_PROPERTY(QString fileTransferUrl READ getFileTransferUrl WRITE setFileTransferUrl NOTIFY fileTransferUrlChanged)
 
 	Q_PROPERTY(bool limeIsSupported READ getLimeIsSupported CONSTANT)
@@ -524,6 +525,9 @@ public:
 
 	bool getChatNotificationsEnabled() const;
 	void setChatNotificationsEnabled(bool status);
+
+	bool getChatReactionsNotificationsEnabled() const;
+	void setChatReactionsNotificationsEnabled(bool status);
 
 	bool getChatNotificationSoundEnabled() const;
 	void setChatNotificationSoundEnabled(bool status);
@@ -875,6 +879,7 @@ signals:
 	void videoConferenceEnabledChanged();
 
 	void chatNotificationsEnabledChanged(bool status);
+	void chatReactionsNotificationsEnabledChanged(bool status);
 	void chatNotificationSoundEnabledChanged(bool status);
 	void chatNotificationSoundPathChanged(const QString &path);
 
@@ -884,7 +889,6 @@ signals:
 	void limeStateChanged(bool state);
 
 	void contactsEnabledChanged(bool status);
-
 	void createEphemeralsChatRoomsChanged();
 
 	// Network. ------------------------------------------------------------------