+++ CHANGELOG.md
@@ -30,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - File permission after auto downloading a file.
 
+### Added
+- Option to mute chat reactions.
+
 ## 5.2.3 - 2024-03-29
 
 ### Fixed